requirements.txt
boto3==1.39.11
openai==1.97.1
python-dotenv==1.1.1